# backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Task Management API
  version: 1.0.0
  description: >-
    A RESTful API for a task management system with JWT authentication,
    role-based access control, task assignment, and analytics.
servers:
  - url: /api/v1
    description: Base API path
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
tags:
  - name: Auth
    description: Registration, login, logout, profile
  - name: Users
    description: User management (admin/manager)
  - name: Tasks
    description: Task CRUD, assignment, analytics
paths:
  /users:
    get:
      summary: List users (admin sees all/filterable, manager sees own team)
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: team
          schema:
            type: string
        - in: query
          name: role
          schema:
            type: string
      responses:
        '200':
          description: List of users
  /users/{id}:
    get:
      summary: Get a single user by id
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User found
        '404':
          description: User not found
    patch:
      summary: Update a user (role/team changes require admin)
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User updated
    delete:
      summary: Delete a user (admin only)
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User deleted
  /tasks:
    post:
      summary: Create a task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                description:
                  type: string
                dueDate:
                  type: string
                  format: date-time
                priority:
                  type: string
                  enum:
                    - low
                    - medium
                    - high
                status:
                  type: string
                  enum:
                    - pending
                    - in-progress
                    - completed
                assignedTo:
                  type: string
                  description: User id
      responses:
        '201':
          description: Task created
    get:
      summary: List tasks with filtering, search, sorting and pagination
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum:
              - pending
              - in-progress
              - completed
        - in: query
          name: priority
          schema:
            type: string
            enum:
              - low
              - medium
              - high
        - in: query
          name: search
          schema:
            type: string
          description: Full-text search across title/description
        - in: query
          name: dueBefore
          schema:
            type: string
            format: date
        - in: query
          name: dueAfter
          schema:
            type: string
            format: date
        - in: query
          name: sortBy
          schema:
            type: string
            default: createdAt
        - in: query
          name: sortOrder
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of tasks visible to the requester
  /tasks/assigned/me:
    get:
      summary: List tasks assigned to the current user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tasks assigned to the requester
  /tasks/analytics:
    get:
      summary: Task completion/pending/overdue statistics (bonus)
      tags:
        - Tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Aggregated statistics scoped to the requester's visibility
  /tasks/{id}:
    get:
      summary: Get a task by id
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task found
        '404':
          description: Task not found
    put:
      summary: Update a task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task updated
    delete:
      summary: Delete a task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task deleted
  /tasks/{id}/assign:
    patch:
      summary: Assign a task to a user (managers limited to their own team)
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - assignedTo
              properties:
                assignedTo:
                  type: string
      responses:
        '200':
          description: Task assigned
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  example: johndoe
                email:
                  type: string
                  example: john@example.com
                password:
                  type: string
                  example: StrongPass1!
      responses:
        '201':
          description: User registered
        '409':
          description: Username or email already exists
  /auth/login:
    post:
      summary: Log in and receive a JWT
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identifier
                - password
              properties:
                identifier:
                  type: string
                  description: Username or email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          returns JWT: null
        '401':
          description: Invalid credentials
  /auth/logout:
    post:
      summary: Log out the current user (invalidates the JWT)
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out
  /auth/me:
    get:
      summary: Get the authenticated user's profile
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile returned
        '401':
          description: Not authenticated